HYD-5466 HYD-5414 HYD-5409 Changes needed to build combined IEEL
Various changes needed to handle the combined IEEL tarball
distribution.

  - ieel job does not have a chroma-bundles subdir
  - set TEST_DISTRIBUTION and friends according to distro
  - EL7 doesn't have kernel-firmware
  - need to make bundle names unique with the distribution
    during bundle installation
  - remove the distribution spec for the lustre-client bundle
    description name
  - init 6 on EL7 reacts so quickly that it shuts down ssh
    sessions before the initiator has time to close it off
    causing the initiator to see a broken connection and
    return an error
    - so give ourselves 2 seconds grace time to shut down
      the session before the rug gets pulled out
 - fix python site-packages paths version hard coding
   - it would be nice to use site.getsitepackages() but that
     doesn't work in virtualenvs
     - https://github.com/pypa/virtualenv/issues/355
 - remove hard-coding of EPEL version
 - remove hard-coding of slave names
 - loosen up the nginx specification in services/test_dependencies
   so that nginx is copied on both EL6 and EL7
 - fix up a previous commit that wanted to print the class a
   failure happened in but forgot to include a %s to print
   the class in

Reviewed-on: http://review.whamcloud.com/18233
Tested-by: Jenkins

+++ chroma-manager/tests/framework/integration/utils/install_client.sh
@@ -36,12 +36,17 @@ yumdownloader lustre-client lustre-client-modules
 # other dependencies and the 'kernel = ' at the beggining of the line.
 KERNEL_VERSION_AND_RELEASE=\"\$(rpm -qpR lustre-client-modules-* | sed -n -e '/^kernel =/s/.* = //p')\";
 
-yumdownloader kernel-\$KERNEL_VERSION_AND_RELEASE kernel-firmware-\$KERNEL_VERSION_AND_RELEASE  # Have to install explicitly to avoid getting kernel-debug in the automatic dependency resolution
+if [[ $(lsb_release -sr) =~ 7.* ]]; then
+    yumdownloader kernel-firmware-\$KERNEL_VERSION_AND_RELEASE
 
-if ! rpm -q \$(rpm -qp kernel-firmware-*.rpm); then
-    rpm -Uvh --oldpackage kernel-firmware-*.rpm
+    if ! rpm -q \$(rpm -qp kernel-firmware-*.rpm); then
+        rpm -Uvh --oldpackage kernel-firmware-*.rpm
+    fi
+    rm kernel-firmware-*.rpm
 fi
-rm kernel-firmware-*.rpm
+
+yumdownloader kernel-\$KERNEL_VERSION_AND_RELEASE # Have to install explicitly to avoid getting kernel-debug in the automatic dependency resolution
+
 if ! rpm -q \$(rpm -qp kernel-*.rpm); then
     rpm -ivh --oldpackage kernel*.rpm
 fi
@@ -53,4 +58,5 @@ yum install -y lustre-client-modules-*.rpm lustre-client-*.rpm
 # Removed and installed a kernel, so need a reboot
 sync
 sync
-init 6"
+nohup bash -c \"sleep 2; init 6\" >/dev/null 2>/dev/null </dev/null & exit 0"
+
